back/microserviciosConArquitectura/pet2pet/requirements.txt

   # PARA INSTALAR LAS DEPENDENCIAS DESDE LA CARPETA RAIZ CON EL ENTORNO ACTIVO EJECUTAR:
    #pip install -r requirements.txt



# Web Framework y Servidor
fastapi==0.104.1
uvicorn==0.24.0

# Base de datos y ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.10

# Autenticación y Seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuración y Validación
python-dotenv==1.0.0
pydantic==2.9.2
pydantic-settings==2.6.0
setuptools==75.2.0

# Manejo de errores y reintentos
tenacity==8.2.3

# Testing
pytest==7.4.3
httpx==0.25.1

# Logging y Monitoreo
python-logging-loki==0.3.1

# Nuevas dependencias recomendadas
email-validator==2.1.0  # Para validación de emails
python-jose[cryptography]==3.3.0  # Asegura que cryptography esté instalado
redis==5.0.1  # Para caché y rate limiting
celery==5.3.6  # Para tareas asíncronas
flower==2.0.1  # Monitoreo de Celery
boto3==1.29.3  # Para AWS S3 (almacenamiento de archivos)
pillow==11.0.0  # Para procesamiento de imágenes
python-magic==0.4.27  # Para verificación de tipos de archivo
requests==2.31.0  # Para llamadas HTTP
aiohttp==3.9.0  # Para llamadas HTTP asíncronas
prometheus-client==0.19.0  # Para métricas
sentry-sdk==1.35.0  # Para monitoreo de errores
bcrypt==4.0.1  # Para hashing de contraseñas
python-multipart==0.0.6 #para procesar datos de formularios enviados a través de solicitudes HTTP con el método POST
aiofiles==24.1.0 #permite trabajar con archivos de manera asíncrona
psutil==6.1.0 #para obtener información sobre los procesos en el sistema y el uso de recursos del sistema
pandas==2.2.3
openptxl==3.1.5

# Herramientas de desarrollo
black==23.11.0  # Formateador de código
flake8==6.1.0  # Linter
mypy==1.7.0  # Verificación de tipos
isort==5.12.0  # Ordenar imports
pytest-cov==4.1.0  # Cobertura de tests
pytest-asyncio==0.21.1  # Para tests asíncronos
faker==20.0.0  # Para generar datos de prueba